libstagefrighthw: Changing TARGET_PRODUCT to QCOM_TARGET_PRODUCT
Using TARGET_PRODUCT to refer to qcom boardnames conflicts with
the customers' definition of the variable (who use it to refer
to the end user visible name for the product). The variable
QCOM_TARGET_PRODUCT has been introduced to refer to the qcom
boardnames. This should be used instead.

Targets tested on: msm8660_surf
Tests: make libstagefrighthw from android root

Change-Id: I20f92a0a9700519d6a3becd128843c05ab9ff83c