Camera: Clear the thumbnail heap during camera release.
Currently the thumbnail heap is cleared after the first
preview frame after snapshot is sent for display. But if
the application issued a release in the middle of snapshot,
then the thumbnail heap is not cleared during release. This
results in parts of relese sequence of the previous instance
to run during the initialization of the next camera instance.

Change-Id: Ied7e209d82170756a6aea9aa681ef913bd4b040a
CRs-Fixed: 284207